Cleaning things up a bit
